# backend/openapi.yml
openapi: 3.0.3
info:
  title: CapeConnect API
  version: 1.1.0
  description: >
    Live backend contract for the current CapeConnect static frontend and
    supporting admin tools. Authentication uses opaque bearer session tokens,
    not JWTs.
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
...
paths:
...
  /api/admin/bootstrap:
    get:
      tags: [Admin]
      summary: Load admin bootstrap data for one operator
      security:
        - bearerAuth: []
      parameters:
        - name: operator
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Admin bootstrap payload
  /api/admin/users/bulk:
    post:
      tags: [Admin]
      summary: Bulk update operator-scoped users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bulk update applied
  /api/admin/tickets/bulk:
    post:
      tags: [Admin]
      summary: Bulk replace operator-scoped tickets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bulk update applied
  /api/admin/wallets/bulk:
    post:
      tags: [Admin]
      summary: Bulk replace operator-scoped wallets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bulk update applied
  /api/admin/prices/global/bulk:
    post:
      tags: [Admin]
      summary: Bulk upsert operator fare products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bulk update applied
  /api/admin/prices/routes/bulk:
    post:
      tags: [Admin]
      summary: Bulk replace operator route pricing
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bulk update applied
  /api/admin/timetables/bulk:
    post:
      tags: [Admin]
      summary: Bulk replace operator timetables
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bulk update applied
  /api/admin/audit:
    get:
      tags: [Admin]
      summary: List audit entries for one operator
      security:
        - bearerAuth: []
      parameters:
        - name: operator
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audit log
    post:
      tags: [Admin]
      summary: Write an audit entry for one operator
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit entry recorded